Support rebuild and unshelve with ephemeral encryption
When an instance is created with ephemeral encryption specified, the
image properties related to the encryption (if present) are taken from
the image and then stored in the instance system metadata. This
includes the secret UUID needed to read the image.

When an instance is snaphotted or shelved, the created image will have
a new secret and its UUID will be kept in the created image's image
properties.

If/when the encrypted snapshot image is used later to create a new
instance or unshelve an existing instance, the new image's image
properties including its secret UUID will be needed to create or
unshelve the instance.

This enables ephemeral encryption related image properties to be
refreshed in the instance system metadata when a new image is used
during a rebuild or unshelve.

Related to blueprint ephemeral-storage-encryption

+++ nova/virt/libvirt/driver.py
@@ -1649,8 +1649,8 @@ class LibvirtDriver(driver.ComputeDriver):
         :param cleanup_instance_dir: If the instance dir should be removed
         :param cleanup_instance_disks: If the instance disks should be removed.
             Also removes ephemeral encryption secrets, if present.
-        :param destroy_secrets: If the cinder volume encryption secrets should
-            be deleted.
+        :param destroy_secrets: If the cinder volume or ephemeral encryption
+            secrets should be deleted.
         """
         # zero the data on backend pmem device
         vpmems = self._get_vpmems(instance)
@@ -1714,7 +1714,7 @@ class LibvirtDriver(driver.ComputeDriver):
             except exception.InstanceNotFound:
                 pass
 
-        if cleanup_instance_disks:
+        if cleanup_instance_disks and destroy_secrets:
             crypto.delete_vtpm_secret(context, instance)
             # Make sure that the instance directory files were successfully
             # deleted before destroying the encryption secrets in the case of
